refactor(): Move Session concern
